Move tests in partial_json.rs to tests/
